Выделен отдельный компонент EnhancedTableToolbar
