+++ src/components/UI/CustomTable/EnhancedTableToolbar.js
@@ -1,0 +1,68 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { lighten, makeStyles, Toolbar, Typography, Tooltip, IconButton } from '@material-ui/core';
+import DeleteIcon from '@material-ui/icons/Delete';
+import FilterListIcon from '@material-ui/icons/FilterList';
+import clsx from 'clsx';
+
+const useToolbarStyles = makeStyles((theme) => ({
+  root: {
+    paddingLeft: theme.spacing(2),
+    paddingRight: theme.spacing(1),
+  },
+  highlight:
+    theme.palette.type === 'light'
+      ? {
+          color: theme.palette.secondary.main,
+          backgroundColor: lighten(theme.palette.secondary.light, 0.85),
+        }
+      : {
+          color: theme.palette.text.primary,
+          backgroundColor: theme.palette.secondary.dark,
+        },
+  title: {
+    flex: '1 1 100%',
+  },
+}));
+
+const EnhancedTableToolbar = ({ numSelected, tableTitle }) => {
+  const classes = useToolbarStyles();
+
+  return (
+    <Toolbar
+      className={clsx(classes.root, {
+        [classes.highlight]: numSelected > 0,
+      })}
+    >
+      {numSelected > 0 ? (
+        <Typography className={classes.title} color='inherit' variant='subtitle1' component='div'>
+          {numSelected} selected
+        </Typography>
+      ) : (
+        <Typography className={classes.title} variant='h6' id='tableTitle' component='div'>
+          {tableTitle}
+        </Typography>
+      )}
+
+      {numSelected > 0 ? (
+        <Tooltip title='Delete'>
+          <IconButton aria-label='delete'>
+            <DeleteIcon />
+          </IconButton>
+        </Tooltip>
+      ) : (
+        <Tooltip title='Filter list'>
+          <IconButton aria-label='filter list'>
+            <FilterListIcon />
+          </IconButton>
+        </Tooltip>
+      )}
+    </Toolbar>
+  );
+};
+
+EnhancedTableToolbar.propTypes = {
+  numSelected: PropTypes.number.isRequired,
+};
+
+export default EnhancedTableToolbar;
